remove institution_reference from project creation form for institutions with funding workflow
currently owning institution project reference is used to refer to the finance code of a grant associated with the project, and not much else.
with the attribution workflow providing a dedicated place for finance codes, this field is now redundant so should not be shown for institutions
opted in to teh new workflow.